move useWaitForTransaction to navbar and wrapped app with token/bounty provider
