+++ components/HeaderNav/HeaderNav.tsx
@@ -6,11 +6,16 @@ import {
   HStack,
   Text,
   Img,
+  toast,
+  useToast,
 } from "@chakra-ui/react";
 import { useRouter } from "next/router";
 import { handleWalletConnect } from "../../utils/walletUtils";
 import Link from "next/link";
-import { useEnsName } from "wagmi";
+import { useEnsName, useWaitForTransaction } from "wagmi";
+import useUser from "../../context/TwaliContext";
+import { useBounty } from "../../context/BountyContext";
+import { useToken } from "../../context/TokenContext";
 
 const HeaderNav = (props) => {
   const whichPage = props.whichPage;
@@ -18,6 +23,11 @@ const HeaderNav = (props) => {
   const setUserData = props.setUserData;
   const userPage = props.userPage;
   const userWallet = props.userWallet;
+  const { ...userState } = useUser();
+  const { ...bountyState } = useBounty();
+  const { tokenName, tokenAmount, calculatedUSD } = useToken();
+  const toast = useToast();
+
   const {
     data: ensData,
     isError: ensError,
@@ -28,6 +38,80 @@ const HeaderNav = (props) => {
   const [isSubmitted, setIsSubmitted] = useState(false);
   const [loaded, setLoaded] = useState(false);
   const router = useRouter();
+  const {
+    data: txData,
+    isError: isTxError,
+    isLoading: txIsLoading,
+  } = useWaitForTransaction({
+    hash: userState.txHash,
+    async onSettled(contractData, error) {
+      if (contractData && !error) {
+        toast({
+          title: "Your bounty was submitted!",
+          description: `${bountyState.contractTitle} is up on the marketplace.`,
+          status: "success",
+          variant: "subtle",
+          duration: 6000,
+          isClosable: true,
+        });
+        let bounty = {
+          ...bountyState,
+          token: tokenName,
+          contractAmount: tokenAmount,
+          convertedAmount: calculatedUSD,
+          userWallet: userState.userWallet,
+          contractOwnerUserName: userState.userName,
+          contractID: bountyState.contractID,
+          contractCreatedOn: Date.now(),
+          contractStatus: "live",
+          attachedFiles: bountyState.attachedFiles,
+        };
+        let URI;
+
+        const isValid = submitSOWToS3(bounty);
+        if (isValid) {
+          isValid.then(async (valid) => {
+            if (valid.status === 200) {
+              URI = await valid.json();
+              await updateSOWToLiveStatus({
+                ...bounty,
+                contractURI: URI,
+              });
+            }
+          });
+        }
+        bountyState.setBounty({});
+      }
+
+      if (error) {
+        toast({
+          title: "Your bounty was not created due to an error!",
+          description: `${error}`,
+          status: "error",
+          variant: "subtle",
+          duration: 5000,
+          isClosable: true,
+        });
+      }
+    },
+  });
+
+  const submitSOWToS3 = async (bounty) => {
+    // post the SOW object to an S3 bucket
+    let res = await fetch("/api/users/postSOW", {
+      method: "POST",
+      body: JSON.stringify({ bounty }),
+    });
+    return res;
+  };
+
+  const updateSOWToLiveStatus = async (bounty) => {
+    let res = await fetch("/api/marketplace/submitBounty", {
+      method: "POST",
+      body: JSON.stringify({ bounty }),
+    });
+    return res;
+  };
 
   return (
     <Flex
@@ -126,7 +210,7 @@ const HeaderNav = (props) => {
                     padding="4px 8px"
                     noOfLines={1}
                   >
-                    {ensData || userWallet}
+                    {txIsLoading ? "tx pending" : ensData || userWallet}
                   </Text>
                 </Flex>
               </HStack>
